app/requirements/base.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
django-environ==0.8.1
psycopg2==2.9.2
django-filter==21.1
redis==4.1.0
celery==5.2.3
flower==1.0.0
kavenegar==1.1.2